t/unit-tests: convert hash to use clar test framework
Adapts the `check_hash_data` function to clar framework by using clar
assertions where necessary. This change follows the consensus to convert
the unit-tests file found in the t/unit-tests folder to clar driven by
Patrick Steinhardt. Test functions were structured as a standalone to
test each individual string and literal case using the refactored
macros.

Mentored-by: Patrick Steinhardt <ps@pks.im>
